+++ docs/dependencies.rst
@@ -41,8 +41,7 @@ Name                                                           Version  License
 `jQuery <http://jquery.com/>`_                                 3.6.0    MIT
 `jQuery Cookie <https://github.com/carhartl/jquery-cookie>`_   1.4.1    MIT
 `jQuery UI <https://jqueryui.com>`_                            1.13.1   MIT
-`Underscore <http://underscorejs.org/>`_                       1.13.2   MIT
-`Codemirror <http://codemirror.net/>`_                         5.65.1   MIT
+`Codemirror <http://codemirror.net/>`_                         6.65.7   MIT
 `floatThead <http://mkoryak.github.io/floatThead/>`_           1.4.0    MIT
 `list.js <http://listjs.com>`_                                 1.2.0    MIT
 `pivottable.js <http://nicolas.kruchten.com/pivottable/>`_     2.0.2    MIT
