+++ explorer/static/js/src/favorites.js
@@ -1,16 +1,17 @@
-let toggle_favorite = function () {
+import $ from 'jquery';
+
+export function toggleFavorite() {
     let queryId = $(this).data('id');
     let favoriteUrl =  $(this).data('url');
     $.post(favoriteUrl, {}, function (data) {
         let is_favorite = data.is_favorite;
         let selector = '.query_favorite_toggle[data-id=' + queryId + ']';
         if (is_favorite) {
-            $(selector).removeClass("glyphicon-heart-empty").addClass("glyphicon-heart");
+            $(selector).removeClass("bi-heart-fill").addClass("bi-heart");
         } else {
-            $(selector).removeClass("glyphicon-heart").addClass("glyphicon-heart-empty");
+            $(selector).removeClass("bi-heart").addClass("bi-heart-fill");
         }
     }.bind(this), 'json').fail(function () {
         alert("error");
     });
 }
-$('.query_favorite_toggle').click(toggle_favorite);
