+++ explorer/static/scss/main.scss
@@ -1,0 +1,7 @@
+@import "~bootstrap/scss/bootstrap";
+
+$bootstrap-icons-font-src: url("~bootstrap-icons/font/fonts/bootstrap-icons.woff2") format("woff2"), url("~bootstrap-icons/font/fonts/bootstrap-icons.woff") format("woff");
+@import "~bootstrap-icons/font/bootstrap-icons";
+
+@import "explorer.css";
+@import "pivot.min.css";
