+++ explorer/templates/explorer/base.html
@@ -1,29 +1,23 @@
 {% load i18n static %}
+{% load vite %}
 
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{% trans "SQL Explorer" %}{% if query %} - {{ query.title }}{% elif title %} - {{ title }}{% endif %}</title>
 
-    <link rel="stylesheet" href="{% static 'explorer/css/codemirror.min.css' %}" />
-    <link rel="stylesheet" href="{% static 'explorer/css/pivot.min.css' %}" />
-    <link rel="stylesheet" href="{% static 'explorer/css/bootstrap.min.css' %}" />
-    <link rel="stylesheet" href="{% static 'explorer/css/explorer.css' %}" />
-    {% block extrastyle %}{% endblock %}
+    {% block style %}
+      {% vite_asset 'scss/main.scss' %}
+    {% endblock style %}
 
-    <script src="{% static 'explorer/js/jquery.min.js' %}"></script>
-    <script src="{% static 'explorer/js/jquery.cookie.min.js' %}"></script>
-
-    <script src="{% static 'explorer/js/jquery-ui.min.js' %}"></script>
     <script type="text/javascript">
-        queryId = "{% firstof query.id 'new' %}";
         csrfCookieName = "{% firstof csrf_cookie_name 'csrftoken' %}";
         csrfCookieHttpOnly = "{% firstof csrf_cookie_httponly False %}" === "True";
+        clientRoute = "{{ request.resolver_match.url_name }}"
     </script>
-    <script src="{% static 'explorer/js/explorer.js' %}"></script>
 </head>
 
 <body>
@@ -79,16 +73,10 @@
             </div>
         </div>
     {% endblock %}
-    <script src="{% static 'explorer/js/underscore-min.js' %}"></script>
-    <script src="{% static 'explorer/js/bootstrap.min.js' %}"></script>
-    <script src="{% static 'explorer/js/list.min.js' %}"></script>
-    <script src="{% static 'explorer/js/codemirror.min.js' %}"></script>
-    <script src="{% static 'explorer/js/sql.min.js' %}"></script>
-    <script src="{% static 'explorer/js/comment.min.js' %}"></script>
-    <script src="{% static 'explorer/js/jquery.floatThead.min.js' %}"></script>
-    <script src="{% static 'explorer/js/pivot.min.js' %}"></script>
-    <script src="{% static 'explorer/js/xlsx.mini.min.js' %}"></script>
-    <script src="{% static 'explorer/js/favorites.js' %}"></script>
+    {% block bottom_script %}
+      {% vite_hmr_client %}
+      {% vite_asset 'js/src/main.js' %}
+    {% endblock bottom_script %}
     {% block sql_explorer_scripts %}{% endblock %}
 </body>
