+++ explorer/templates/explorer/fullscreen.html
@@ -6,9 +6,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{% trans "SQL Explorer" %}{% if query %} - {{ query.title }}{% elif title %} - {{ title }}{% endif %}</title>
-
-    <link rel="stylesheet" href="{% static 'explorer/css/bootstrap.min.css' %}" />
-    <link rel="stylesheet" href="{% static 'explorer/css/explorer.css' %}" />
 </head>
 
 <body>
