+++ explorer/templates/explorer/query_favorite_button.html
@@ -1,6 +1,6 @@
 {% if is_favorite %}
-    <i class="glyphicon glyphicon-heart text-danger {{ extra_classes }}" data-id="{{ query_id }}" data-url="{% url 'query_favorite' query_id%}"></i>
+    <i class="bi-heart-fill text-danger {{ extra_classes }}" data-id="{{ query_id }}" data-url="{% url 'query_favorite' query_id%}"></i>
 {% else %}
-    <i class="glyphicon glyphicon-heart-empty text-danger {{ extra_classes }}" data-id="{{ query_id }}" data-url="{% url 'query_favorite' query_id%}"></i>
+    <i class="bi-heart text-danger {{ extra_classes }}" data-id="{{ query_id }}" data-url="{% url 'query_favorite' query_id%}"></i>
 {% endif %}
