+++ package.json
@@ -1,0 +1,46 @@
+{
+  "name": "django-sql-explorer",
+  "description": "Django SQL Explorer",
+  "scripts": {
+    "dev": "npx vite --config explorer/static/config/vite.config.js",
+    "build": "npx vite build --config explorer/static/config/vite.config.js",
+    "preview": "npx vite preview --config explorer/static/config/vite.config.js",
+    "lint-sass": "npx stylelint --config explorer/static/config/stylelint.config.js ./src/scss/",
+    "format-sass": "npx stylelint --config explorer/static/config/stylelint.config.js --fix 'src/**/*.{css,scss}'",
+    "lint-js": "npx eslint --config explorer/static/config/eslint.config.js ./src/js",
+    "format-js": "npx eslint --config explorer/static/config/eslint.config.js --fix ./src/js"
+  },
+  "devDependencies": {
+    "@popperjs/core": "^2.11.6",
+    "bootstrap": "^5.0.1",
+    "eslint": "^8.48.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "sass": "~1.69.0",
+    "stylelint": "^15.6.2",
+    "stylelint-config-recess-order": "^4.0.0",
+    "stylelint-config-standard-scss": "^11.0.0",
+    "stylelint-scss": "^5.0.0",
+    "vite": "^5.0.1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/chrisclark/django-sql-explorer.git"
+  },
+  "author": "Brent O'Connor",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/epicserve/django-base-site/issues"
+  },
+  "homepage": "https://github.com/epicserve/django-base-site#readme",
+  "dependencies": {
+    "bootstrap-icons": "^1.11.2",
+    "codemirror": "^6.0.1",
+    "cookiejs": "^2.1.3",
+    "floatthead": "^2.2.5",
+    "jquery": "^3.7.1",
+    "jquery-ui": "^1.13.2",
+    "list.js": "^2.3.1",
+    "pivottable": "^2.23.0"
+  }
+}
